Remove duplicated code from new files
