+++ client/render/SDLImageLoader.h
@@ -1,0 +1,54 @@
+/*
+ * CAnimation.h, part of VCMI engine
+ *
+ * Authors: listed in file AUTHORS in main folder
+ *
+ * License: GNU General Public License v2.0 or later
+ * Full text of license available in license.txt file, in main folder
+ *
+ */
+#pragma once
+
+#include "../../lib/GameConstants.h"
+
+#ifdef IN
+#undef IN
+#endif
+
+#ifdef OUT
+#undef OUT
+#endif
+
+VCMI_LIB_NAMESPACE_BEGIN
+
+class JsonNode;
+class Rect;
+class Point;
+
+VCMI_LIB_NAMESPACE_END
+
+struct SDL_Surface;
+struct SDL_Color;
+class CDefFile;
+class ColorFilter;
+
+
+class SDLImageLoader
+{
+	SDLImage * image;
+	ui8 * lineStart;
+	ui8 * position;
+public:
+	//load size raw pixels from data
+	inline void Load(size_t size, const ui8 * data);
+	//set size pixels to color
+	inline void Load(size_t size, ui8 color=0);
+	inline void EndLine();
+	//init image with these sizes and palette
+	inline void init(Point SpriteSize, Point Margins, Point FullSize, SDL_Color *pal);
+
+	SDLImageLoader(SDLImage * Img);
+	~SDLImageLoader();
+};
+
+
